feat(hero-section): implement animated chat interface for user interaction to enhance user experience refactor(hero-section): restructure layout to improve responsiveness and visual appeal chore(human-loop-section): remove obsolete HumanLoopSection component to streamline codebase fix(page): remove HumanLoopSection import and usage to prevent errors and clean up the home page
