+++ app/(logged-out)/home/page.tsx
@@ -7,7 +7,6 @@ import { CTASection } from './components/cta-section';
 import { FAQSection } from './components/faq-section';
 import { FeaturesBentoGrid } from './components/features-bento-grid';
 import { HeroSection } from './components/hero-section';
-import { HumanLoopSection } from './components/human-loop-section';
 import { PrivateAlphaModal } from './components/private-alpha-modal';
 import { ReviewsMarquee } from './components/reviews-marquee';
 
@@ -44,9 +43,6 @@ export default function HomePage() {
 
       {/* Bento Grid Features */}
       <FeaturesBentoGrid />
-
-      {/* Human in the Loop Section */}
-      <HumanLoopSection />
 
       {/* Reviews Section */}
       <ReviewsMarquee />
